structuring tests and functions
